third_party/blink/web_tests/wpt_internal/app-history/app-history-entry/entries-after-javascript-url-navigation.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    let start_key = i.contentWindow.appHistory.current.key;
    let start_url = i.contentWindow.appHistory.current.url;
    let start_id = i.contentWindow.appHistory.current.id;
    i.onload = t.step_func_done(() => {
      assert_equals(i.contentWindow.appHistory.entries().length, 1);
      let js_url_entry = i.contentWindow.appHistory.current;
      assert_equals(js_url_entry.key, start_key);
      assert_equals(js_url_entry.url, start_url);
      assert_not_equals(js_url_entry.id, start_id);
    });
    i.src = "javascript:'new content'";
  });
}, "AppHistory behavior after navigation to a javascript: url");
</script>
